refactor: ResultView 테스트 분리
